+++ pmlc/rt/level_zero/level_zero_invocation.h
@@ -11,14 +11,14 @@ namespace pmlc::rt::level_zero {
 class LevelZeroInvocation;
 class LevelZeroEvent;
 
-/// Class encapsulating level zero memory buffer allocated on device.
+/// Class encapsulating LevelZero memory buffer allocated on device.
 class LevelZeroMemory {
 public:
   LevelZeroMemory(void *buffer, size_t bytes, ze_context_handle_t context)
       : buffer(buffer), bytes(bytes), context(context) {}
   ~LevelZeroMemory() { lzu::free_memory(context, buffer); }
 
-  /// Returns OpenCL buffer.
+  /// Returns allocated buffer.
   void *getBuffer() { return buffer; }
   /// Returns size of buffer in bytes.
   size_t size() { return bytes; }
@@ -39,13 +39,10 @@ private:
   ze_context_handle_t context;
 };
 
-/// OpenCL kernel with additional state information:
-/// name and event dependencies.
-/// Serves as one time incremental object, `enqueue()` should be
-/// called only once and does not reset it's internal state.
+/// LevelZero kernel wrapper
 class LevelZeroKernel {
 public:
-  /// Constructs kernel declared with `name` from compiled `program`.
+  /// Constructs kernel declared with `name` from compiled `module`.
   LevelZeroKernel(ze_module_handle_t module, std::string name);
   ~LevelZeroKernel();
 
@@ -55,7 +52,8 @@ public:
   void setArg(unsigned idx, LevelZeroMemory *memory);
   /// Enqueues wrapped kernel on specified command queue `queue` with
   /// `gws` global work size and `lws` local work size.
-  /// Returns OpenCL event tracking execution of kernel execution.
+  /// The gws and lws shall be converted for LevelZero API.
+  /// Returns LevelZero event tracking execution of kernel execution.
   void enqueue(ze_command_list_handle_t list, ze_group_count_t gws,
                ze_group_count_t lws, ze_event_handle_t &resultE);
   /// Returns name of this kernel.
@@ -70,24 +68,24 @@ private:
   std::vector<ze_event_handle_t> dependencies;
 };
 
-/// Kind of asynchronous operation executed on OpenCL device.
+/// Kind of asynchronous operation executed on LevelZero device.
 enum class LevelZeroActionKind { Barrier, Kernel, Read, Write };
 
-/// Class encapsulating OpenCL event that describes status of
+/// Class encapsulating LevelZero event that describes status of
 /// operation that produced it and serves for ordering operations.
 class LevelZeroEvent {
 public:
   LevelZeroEvent(ze_event_handle_t event, LevelZeroActionKind kind,
                  std::string name);
 
-  /// Returns OpenCL event object.
+  /// Returns LevelZero event object.
   ze_event_handle_t getEvent() const { return event; }
   /// Returns kind of operation that this event describes.
   LevelZeroActionKind getKind() const { return kind; }
   /// Returns name of operation that this event describes.
   const std::string &getName() const { return name; }
 
-  /// Blocks execution untill all `events` have finished executing.
+  /// Blocks execution until all `events` have finished executing.
   static void wait(const std::vector<LevelZeroEvent *> &events);
 
 private:
@@ -96,18 +94,18 @@ private:
   std::string name;
 };
 
-// OpenCLInvocation encapsulates a particular run of a network on a OpenCL
+// LevelZeroInvocation encapsulates a particular run of a network on a LevelZero
 // device. It's instantiated and managed from the JITted network code, using
-// callbacks in opencl_wrappers.cc.
+// callbacks in level_zero_wrappers.cc.
 class LevelZeroInvocation {
 public:
   explicit LevelZeroInvocation(LevelZeroDevice *device);
   ~LevelZeroInvocation();
 
-  /// Allocates memory on OpenCL device with specified size in bytes.
+  /// Allocates memory on LevelZero device with specified size in bytes.
   LevelZeroMemory *allocateMemory(size_t bytes);
   /// Releases memory obtained from `allocateMemory` call.
-  /// Any further use of `memory` is invalid.
+  /// But as kernel may not really run here, may just do record.
   void deallocateMemory(LevelZeroMemory *memory);
   /// Enqueues operation to read data from `src` memory into `dst` pointer.
   /// Read will start executing only after operations connected to `deps`
@@ -121,20 +119,15 @@ public:
   /// Returns event describing status of this operation.
   LevelZeroEvent *enqueueWrite(LevelZeroMemory *dst, void *src,
                                const std::vector<LevelZeroEvent *> &deps);
-  /// Creates OpenCL kernel from SPIRV binary.
+  /// Creates LevelZero kernel from SPIRV binary.
   LevelZeroKernel *createKernelFromIL(char *data, size_t bytes,
                                       const char *name);
   /// Enqueues kernel execution on device with specified global and local work
-  /// sizes. Kernel object is destroyed after this function finishes. Returns
+  /// sizes. Kernel object needs to be destroyed manually. Returns
   /// event describing status of kernel execution.
   LevelZeroEvent *enqueueKernel(LevelZeroKernel *kernel, ze_group_count_t gws,
                                 ze_group_count_t lws);
-  /// Enqueues barrier operation on device. Barrier serves as synchronization
-  /// primitive that ensures all operations enqueued after it finishes will
-  /// start executing only after all operations before have finished.
-  /// Barrier goes into effect after dependant events `deps` are finished
-  /// or immiediately if `deps` is empty.
-  /// Returns event describing status of this barrier operation.
+  /// Not used now
   LevelZeroEvent *enqueueBarrier(const std::vector<LevelZeroEvent *> &deps);
   /// Forces starting execution of previosuly enqueued operations that aren't
   /// blocked by any dependencies.
