+++ pmlc/rt/level_zero/tests/BUILD
@@ -5,5 +5,5 @@ load("//pmlc:lit.bzl", "glob_lit_tests")
 package(default_visibility = ["//visibility:public"])
 
 glob_lit_tests(
-  data = ["//pmlc/rt/opencl:clprobe"],
+  data = ["//pmlc/rt/level_zero:level_zero_probe"],
 )
