+++ pmlc/rt/level_zero/tests/lit.local.cfg
@@ -2,15 +2,15 @@ import os
 import platform
 import subprocess
 
-clprobe = os.path.join('pmlc/rt/opencl/clprobe')
+levelZeroProbe = os.path.join('pmlc/rt/level_zero/level_zero_probe')
 if platform.system() == 'Windows':
-    clprobe += '.exe'
+    levelZeroProbe += '.exe'
 
-if not os.path.exists(clprobe):
+if not os.path.exists(levelZeroProbe):
     config.unsupported = True
 else:
-    proc = subprocess.run(clprobe, capture_output=True, text=True)
+    proc = subprocess.run(levelZeroProbe, capture_output=True, text=True)
     if proc.returncode:
         config.unsupported = True
-    elif "No supported OpenCL devices available" in proc.stdout:
+    elif "No supported LevelZero devices available" in proc.stdout:
         config.unsupported = True
